INTYGFV-15388: Refactored
