+++ apps/rehabstod/src/components/Table/filter/print/PrintFilterTitle.tsx
@@ -1,0 +1,3 @@
+export function PrintFilterTitle({ title }: { title: string }) {
+  return <p className="font-bold">{title}:</p>
+}
