+++ apps/rehabstod/src/components/Table/filter/utils/getMultipleSelectPlaceholder.ts
@@ -1,4 +1,4 @@
-import { OccupationType, RekoStatusType } from '../../../schemas/sickLeaveSchema'
+import { OccupationType, RekoStatusType } from '../../../../schemas/sickLeaveSchema'
 
 export const getMultipleSelectPlaceholder = (selected: string[], options: OccupationType[] | RekoStatusType[]) => {
   if (selected.length === 0 || !selected) {
