+++ apps/rehabstod/src/pages/CurrentSickLeaves/components/Filters.tsx
@@ -8,7 +8,7 @@ import { DoctorFilter } from '../../../components/Table/filter/DoctorFilter'
 import { RangeFilter } from '../../../components/Table/filter/RangeFilter'
 import { TimePeriodFilter } from '../../../components/Table/filter/TimePeriodFilter'
 import { MultipleSelectFilterOption } from '../../../components/Table/filter/MultipleSelectFilterOption'
-import { getMultipleSelectPlaceholder } from '../utils/getMultipleSelectPlaceholder'
+import { getMultipleSelectPlaceholder } from '../../../components/Table/filter/utils/getMultipleSelectPlaceholder'
 import { SelectFilter } from '../../../components/Table/filter/SelectFilter'
 import { TextSearchFilter } from '../../../components/Table/filter/TextSearchFilter'
 import { TableFilter } from '../../../components/Table/TableFilter'
