+++ apps/rehabstod/src/pages/LUCertificates/LUCertificatesFilters.tsx
@@ -5,7 +5,7 @@ import { RangeFilter } from '../../components/Table/filter/RangeFilter'
 import { useAppSelector } from '../../store/hooks'
 import { TextSearchFilter } from '../../components/Table/filter/TextSearchFilter'
 import { SelectFilter } from '../../components/Table/filter/SelectFilter'
-import { getMultipleSelectPlaceholder } from '../CurrentSickLeaves/utils/getMultipleSelectPlaceholder'
+import { getMultipleSelectPlaceholder } from '../../components/Table/filter/utils/getMultipleSelectPlaceholder'
 import { MultipleSelectFilterOption } from '../../components/Table/filter/MultipleSelectFilterOption'
 import { resetFilters, updateFilter } from '../../store/slices/luCertificates.slice'
 import { useGetDoctorsForLUCertificatesQuery, useGetPopulatedFiltersQuery, useGetUserQuery } from '../../store/api'
